requirements.txt
boto3~=1.16.3
Pillow~=7.2.0
pylint==2.6.0
